pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crucible-parsers"
version = "0.1.0"
description = "Development parsers for Crucible datasets"
readme = "README.md"
license = {text = "BSD"}
requires-python = ">=3.8"

dependencies = [
    "nano-crucible",
]

[project.entry-points."crucible.parsers"]
matensemble-manager = "crucible_parsers.matensemble:MatEnsembleManagerParser"
matensemble-run     = "crucible_parsers.matensemble:MatEnsembleRunParser"
phasefield-manager  = "crucible_parsers.matensemble:PhaseFieldManagerParser"
phasefield-run      = "crucible_parsers.matensemble:PhaseFieldRunParser"

[tool.setuptools]
packages = ["crucible_parsers", "crucible_parsers.matensemble"]
